Add unit and component testing to project
